requirements.txt
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
openai
pydantic